Add support for custom A11yID in TableViewIllustratedItem
Having a custom accessibilityIdentifier for the cell set through the TableViewItem enables easier egtests and better accessibility without having to set the ID in cellForRowAtIndexPath.

A use case for this is RecentTabs, which will use the same TableViewIllustratedItem for all the empty states, with different ItemTypes. Without this modification to TVIllustratedItem, RecentTabs would have to manually set the a11yID of the cell for every itemType.

Bug: 1098328
Change-Id: I341692643556a42984c843594a9be8d24b139762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2313297
Commit-Queue: Maxime Charland <mcharland@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791019}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b41a4be70dffbab45469a63bc02c896384c97472