[SH iOS] Introduce RegisterEditMenuItem
This CL introduces a new util method, RegisterEditMenuItem, for adding
items to the global Edit Menu singleton. This is needed because the
classes that add to the Edit Menu overwrite the contents of the custom
actions in the singleton.

I also found what I believe to be a small metrics logging error in
LocationBarVC, and updated it to be consistent with OmniboxVC.

Context: Fixing the overwrite behavior described above enables adding
a new Edit Menu action to the content area. This refactor was split off
from:
https://chromium-review.googlesource.com/c/chromium/src/+/2285328

Bug: 1099268
Change-Id: Ia8fad48aad0d047551439faa95c39374a2a8e156
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2313477
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791217}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7955087a0f773aef5c287cfd6c554adefc0f4e35