[iOS] Updates link label to texts view in welcome screen
based on: https://crrev.com/c/2300383
This CL updates the link label used to display the TOS string to an UITextView.
A sublcass is needed to override the |canBecomeFirstResponder| method.

On iOS 13,we can't override |canBecomeFirstResponder| or the whole
string will responds to a tap.

On iOS 12, the whole string responds to a tap if we use this current
implementation.

Bug: 1102996
Change-Id: Id9c9e3089d07a056c365043e1d60b1b558586ccf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2302900
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791755}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf155c5ce59f8b649ba6ed14770803e7b0dc6adc