[iOS] Improve readability of UserSigninViewController.
It was difficult to understand the current constraint layout of the
"Turn on Sync" page due to the inconsistencies in the view manipulation.
This patch attempts to make the layout and styling of the views and
buttons more straightforward.

Bug: 1105853
Change-Id: I0c0bca5aa1be5200d3be414804976efc8ed7ef22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2312556
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#791778}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45c1a3f18218914254a4785578e06d38bd7a48f7