Add a completion handler call back in openSelectedTabInMode
completion should be called in every branch of the function, and it was not in this one,
which is a bug. Separated out from CL: https://chromium-review.googlesource.com/c/chromium/src/+/2236243
so that in case of reverting, this change can stay.

Change-Id: If1cf0522b492e8c240253e89432f16d21916da70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320133
Commit-Queue: Andy Lu <andyhylu@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#792327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 377c904cc698ceff7aafa58ff48e172cc3b6cc61