Chrome Empties: Fix testCachePositionIsResetWhenNodeIsPermanentAndEmpty
The testCachePositionIsResetWhenNodeIsPermanentAndEmpty test in BookmarksTestCase fails when the illustrated-empty-states flag is enabled because while the functionality remains the same, an assumption about the view's contents is different because of Chrome Empties.

The test moves into Mobile Bookmarks, closes Bookmarks, removes all the bookmarks, then reopens Bookmarks to make sure the root level of Bookmarks is opened.
While this is still true, the test checks the existence of "Mobile Bookmarks" in the TableView to see if the views is currently displaying the root level.
With Chrome Empties, the test will check the absence of a Back button, which is the case with or without Chrome Empties.

Bug: 1098328
Change-Id: I52c2c18a5b489febb0ba6e7e76be4c1577de1eac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327230
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Maxime Charland <mcharland@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#793176}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38bf89616bf01027cec4b404ba3b28727a51f8c8