[sync] Expose trusted vault recoverability state in SyncUserSettings
This is intended to be used as an indication to the UI that the user
should take action to improve recoverability of the sync-ed data.

This patch introduces basic plumbing in TrustedVaultClient with TODOs
to address platform-specific implementations in later patches.
Meanwhile, StandaloneTrustedVaultClient is extended with test-only
APIs for integration tests.

Change-Id: Ic255e502d9e33b145272903dc2b7c3a7ec2a1cd8
Bug: 1081649
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329751
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#793231}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ac641dfa5f849cfa9fc0fdae41b744377969eb4