[iOS CBCM] Add support for the CloudReportingEnabled policy
Adds support for the CloudReportingEnabled policy, which controls
whether enterprise reports are sent to the device management server.

This policy is a privacy requirement of launching CBCM. For now, it is
declared as a Future policy for iOS until iOS CBCM is launched.

Bug: 1066495
Change-Id: Ic4d175df8a52dabb841d905eaa7fa76e1a9c13bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332254
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Owen Min <zmin@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#793771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b32bd198649023b4c67145507a9be9f4b5376535