[iOS] Trigger blocking UI while clearing browsing data.
The Clear Browsing Data flow uses the ChromeActivityOverlayCoordinator to
block user interactions in the window in which it was triggered. To
extend this to multiple windows, this CL adds an option to the activity
coordinator to block all windows, implemented by triggering the blocking
UI.

that the blocking UI is shown in the other window.

TEST: Opened two windows, opened History, chose Clear Browsing Data, saw
Bug: 1045580
Change-Id: I8466cd8e67ddedce13c6db6580cd0d57b76b120d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339963
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#795468}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ae5fd76a65044b8e673be15cfcec2ec5a37d46c