Reland: [iOS] Use a persistent cookie store for Safe Browsing
This makes SafeBrowsingService use a persistent cookie store,
and clears this cookie store whenever cookies are cleared for
"all time". This makes the behavior on iOS consistent with
the behavior of Safe Browsing on other platforms.

This is a re-land of http://crrev.com/c/2291694 with a fix
for SafeBrowsingServiceTest.PersistentCookies.

Bug: 1103219
Change-Id: I61dc96bdf8a77c3336bca20878cfc2cda76ee990
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2346729
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#796619}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 028082e315cf7eb4aad6957912b67e6279f6fedf