[ios] UI Polish Omnibox: add animation to DSE favicon.
Current behaviour:
When focusing the Omnibox, the icon on the far left (G icon, Favicon,
e.g.) has no animation.

Intended behaviour:
The icon on the far left should have a short animation starting from an
opacity of 0 and a scale of 0 towards an opacity of 100 and a scale of 1.

Bug: none.
Change-Id: I06a90707db4e496fe1aef032e2ce110f392e3d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339531
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Nazerke Kalidolda <nazerke@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#796931}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 281486c28e8370d85d6ffd5df20c3c36b4835438