[iOS] Remove TabSwitcher protocol
The protocol was here for compatibility reason with the old TabSwitcher.
This CL removes it and moves its methods in the TabGridCoordinator and
rename its methods.
It also renames its delegate to TabGridCoordinatorDelegate.
It removes the parameter id<TabSwitcher> when it is no longer necessary.
It removes the reference to ViewControllerSwapping.

Bug: none
Change-Id: I4d4504d51f0c7560a13296613dfd72ba68aa116c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339974
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#797207}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ead4276bccf17802499bb8c62e70c2d93c091bc3