[ios] Fixed Multi-Window size issue on iPhone landscape
In MW mode, when on iPhones which isn't MW, the frame for a
window is set multiple times by iOS private methods, but some
of those do not take into account screen rotation and can end up
passing in a frame that is in portrait sizes on a landscape device.
When that happens, the frame is centered, hence the origin
contains a negative coordinate.  This CL skips setting the
frame on the ChromeOverlayWindow's parent when it has a negative
x (or y, in case the bug can be reproduced the inverse way).
It applies to iOS 13 and 14.

Bug: 1113187
Change-Id: I38b9f9f475502d5123ab381dbde01296416b22d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359137
Commit-Queue: David Jean <djean@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799037}
GitOrigin-RevId: b352a52d9c933d17bb22a313cd4e019537345e87