[iOS] Fix Tab Grid Recent Tabs Open in New Tab
The issue:
Triggering "Open in New Tab" from the tab grid recent tabs' context menu
did not navigate the user to that tab. The tab count was incremented,
but it was hard to tell from the UI that something happened.

The fix:
Navigate the user to the new tab.

Other fix:
- Added a similar solution for incognito. The issue didn't happen in
  incognito, but I found it awkward to not have a similar approach.

Bug: 1093302
Change-Id: I0513a1195b768bfaa802788443f411b8a24879c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359557
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799130}
GitOrigin-RevId: 3a42ed0d7bd56a8de9102d3a122bfe2011d8718d