[iOS] Use a simpler grey matcher for web view
For some reason chrome_test_util::WebViewMatcher is flaky in these ones.
Could be related to the keyboard being in screen.

The flakiness appears to be related to iOS 13.4. Maybe a recent update
to the bots or a version upgrade of earl grey.

Using a simpler matcher seems to fix the issue.

Fixed: 1116043, 1116734, 1114899, 1111076, 1115321, 1111258, 1115282
Change-Id: I9172417a20d5b8e09dc47cd986256fc0a8cb46ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362892
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799544}
GitOrigin-RevId: ed3fbbd4f1827185819a87a596e73c51e4206e2a