[ios] Replace SnapshotCacheFactory with SnapshotBrowserAgent
Create SnapshotBrowserAgent, whose lifetime is tied to the Browser.
Remove SnapshotCacheFactory, whose lifetime is tied to the BrowserState
(profile). This allows separate windows to have their own folder so that
session data can be purged.

To achieve this:
1) Add a uniqueIdentifier to the SnapshotCache, which is used in the
filepath where snapshots are stored on disk.
2) Set the UISceneSession's persistentIdentifier as the SnapshotCache's
uniqueIdentifier.
3) Plumb the SnapshotCache associated with a Browser to all the
SnapshotTabHelpers and SnapshotGenerators that are associated with
WebStates in the WebStateList in the Browser.
4) Replace all uses of SnapshotCacheFactory with SnapshotBrowserAgent.

Bug: 1110958, 1087843
Change-Id: I1aba5f3b61ae7835658887807b329866bf08ce20
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2348293
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#799678}
GitOrigin-RevId: 95172de18f7e8e820872db6096142a7dccb6c6c7