[iOS CBCM] Enable CBCM at 100% on Beta
Enables CBCM at 100% on Beta on iOS (in addition to the existing
command-line switch, which can enable CBCM in any channel). To
paraphrase the comment in the code, CBCM must either be enabled
at 100% or 0% because the Finch infrastructure isn't ready by the
time CBCM must decide whether to initialize.

Because of additional requirements explained in the following 1-pager
(Google only), CBCM is required on Beta, which means CBCM will be
enabled at 100% on Beta and 0% on all other channels:
go/ios-cbcm-rollout-1pager.

Bug: 1066495
Change-Id: I1449f5314ea592297d79ec7acbbb0991d31bed5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2364912
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: sebsg <sebsg@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800330}
GitOrigin-RevId: 0708646ab6180b66abdb3c88860cf7cd11700036