Chrome Empties: Fix bookmarks title bar bug
This addresses a bug where the title and search bar overlapped in the
NavigationItem in rare cases. Making the call to readd the search bar
before the call to update the title style fixes the behavior, though the
title style now won't change until the user interacts with the screen.

It's hard to say exactly *why* this fixes the issue without some more
knowledge of how UINavigationItem works internally, which we obviously
don't have, but my hunch (and the reason I tried this approach) is that
changing the title style triggers a re-layout of some kind.

Bug: 1116407
Change-Id: I774a1fc9a77a71c01626ddef4dc2586575318a5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367974
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800365}
GitOrigin-RevId: b7858e97429965a33ba0b685474c91f15e70c1eb