[Empties] Reload content before calculating bookmark promo offset.
In some cases the model and table were not in sync, which caused a
crash when trying to get the visibleCells of the table.

The reload should not be too expensive since we only call it when
we're on the root bookmark page.

Bug: 1116408
Change-Id: I41e3f1c6d09ca5b92f2bbb0494d426e04f2db539
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2370605
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: sebsg <sebsg@chromium.org>
Auto-Submit: sebsg <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801008}
GitOrigin-RevId: 9978b8f71fb7550728a986c6c9e9d2cefe6e03e1