[iOS] Prevent re-entrance in omnibox edit menu clipboard
As seen in the attached bug, sometimes accessing the pasteboard from
the PasteboardChangeNotification callback causes the notification to be
posted again, causing an infinite recursion and crash. The bug is most
likely on the Apple end, but without reproduction steps, I'm not sure
if they will respond. Nevertheless, we can work around this by only
updating in response to the notification once at a time. Then the second
notification will be ignored.

Bug: 1049053
Change-Id: If15c8d6b17d39b927ac4d21a643de28fa3851a2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360021
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801366}
GitOrigin-RevId: d76bdbba90f25c5a985e81b089808023360d614b