[iOS CBCM] Fix cloud reporting when tested via experimental settings
The CloudReportingEnabled policy controls whether cloud reports are sent
to the Admin Console. It is still an experimental policy on iOS. When
testing the CBCM flow via experimental settings, that policy was never
allowed in the list of experimental policies, so it was ignored. That
prevented testing cloud reporting.

This change simply adds the CloudReportingEnabled policy to the list of
allowed experimental policies if a CBCM enrollment token is provided via
the experimental settings.

This change also clarifies comments and variable names around policy
about:flags to make it easier to understand what's going on.

Bug: 1122005
Change-Id: Iec6ecf7b4ef7a2a8bed60dbd82b73a474a70cd8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2377896
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#802276}
GitOrigin-RevId: 25e9bd7890a6467b1058bc16c2a503bf0846a233