[iOS] Remove CaptivePortalMetricsTabHelper
This tab helper was used to analyze potential improvements for
detecting users in bad captive portal states. However, it is now no
longer being used. Removing it will clean up captive portal and
application code to enable easier changes to captive portal
code in the future if needed.

This CL is referenced from a new README file so that it can more easily
be discovered in the event that this functionality is found to be useful
in the future.

Bug: 1072565
Change-Id: I0aea60632587780150a0d378438a082824d184fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2372901
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#802455}
GitOrigin-RevId: c5304d7b5d3f92873d33c313b8c856ca6e8e0900