[iOS] Tentative fix for crash
This CL fixes a crash where the navigationItem associated with a
BackForwardListItem is nil on a back/forward navigation, leading
to a null pointer crash.
If the associated item is nil, the code path is now the same as a
non-back/forwarad navigation.

Fixed: 1121950
Change-Id: Ic4b66af2dc9aa368ff046d66095cf090591eed57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2379724
Commit-Queue: Eugene But <eugenebut@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#802824}
GitOrigin-RevId: 52d159212b680fad216e0a7e810e5d2abba168db