Make the thumb strip coordinator a child of tab grid coordinator.
Currently the thumb strip coordinator is owned by the browser
coordinator, which makes it simple to set up references to the tab strip
and toolbar controllers, for example, that are needed to animate the
transition from tab strip to thumb strip. However, the thumb strip will
need a reference to the tab grid view controller in order to propagate
the layout transition from thumb strip to tab grid. This means it's best
for the thumb strip coordinator to be owned by the tab grid coordinator.

- Create the ThumbStripAttacher protocol, implemented by the browser
view controller. It defines an interface to set the pan handler
reference and set up the thumb strip.
- Create two references inside the tab grid coordinator (regular and
incognito thumb strip attachers) that are set up after starting the
thumb strip coordinator.

Change-Id: Ic1d5706c1a9110e4387b5c1a3f8a7d0c9d88cbb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340911
Commit-Queue: Roberto Moura <mouraroberto@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#804797}
GitOrigin-RevId: bc0449ab358d0b7089934c0d0f05492352cba5f6