[SyncInvalidations] Plumb info through DeviceInfoSyncClient.
Before this CL, LocalDeviceInfoProvider was observing SyncInvalidationsService and caching state internally.
After this CL, it instead queries the same information on demand through DeviceInfoSyncClient.

This change is needed because with the current implementation, it's not possible to resend device info when FCM token or interested data types change.

Bug: 1102336
Change-Id: Ic18f164014b899daa38b1570b2a984805365f384
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2396142
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Paula Vidas <paulavidas@google.com>
Cr-Commit-Position: refs/heads/master@{#804875}
GitOrigin-RevId: 0fc91601fa2b42afea057687c7a41ae832ce1d7d