[iOS] Don't donate intents at startup
According to the Apple doc below, donating intents is a way for apps to
tell Siri "The user just performed this action inside my app". This is
used by Siri to give better results to the user when they search in
Spotlight, or to give suggestions to the user on the homescreen, such
as "Do you wish to Search in Chrome?". According to the doc, donating
intents should only happen immediately after the user performs the
related action inside the app.

Also, donating intents is not needed to register shortcut actions. That
happens automatically via the intent definition file.

For those two reasons, there is no need to donate the SearchInChrome
intent during startup, so this CL removes that.

https://developer.apple.com/documentation/sirikit/donating_shortcuts

Bug: 1125200
Change-Id: I93c1e71ced03b70d0849ce27390ccbfe9821b4f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2401761
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#805728}
GitOrigin-RevId: 9300797ac49e53f3292ec5f6c6466b607660623c