Merge URLRequestJobFactoryImpl into URLRequestJobFactory.
There are only four URLRequestJobFactory implementations, and all but
URLRequestJobFactoryImpl are test-only. This CL removes one of them,
one is for FTP, which is deprecated. The last is for Cronet, which may
take some effort to unwrangle.

Ideally we'll eventually remove URLRequestJobFactory as well, but until
then, this simplifies things a bit.

Bug: 354158
Change-Id: I7e644ce6690692adc1231cf3a1464743ced073fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2394155
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806313}
GitOrigin-RevId: a9606a0ebf73739adebeb360a3fd2bb42b8fe534