[ios] Add metric to verify storing prefs::kLastSessionExitedCleanly
Chrome relies on prefs::kLastSessionExitedCleanly pref to verify that
browsing session has ended exited cleanly as oppose to app crash.
The assumption is that prefs::kLastSessionExitedCleanly pref is always
flushed and crash detection logic won't work correctly if assumption is
not correct.

This CL adds a temporary Stability.iOS.UTE.MobileSessionAppState
histogram to verify the assumption. The value of the histogram should
always be ActiveUTE or ActiveXTE. Any other value indicates that
prefs were not properly flushed.

Bug: None
Change-Id: I09d407d645f05a41179a9e1faee54f0cf294ef40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2399769
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Commit-Position: refs/heads/master@{#806383}
GitOrigin-RevId: 44c96eac30b0f2ffbd7d07ecd2b3f05f423d2ef1