[iOS] Add histograms to track OverscrollActions memory usage
Discussions on the bug show that OverscrollActionsController's high
memory usage is coming from a large number of small allocations. The
method in question is the initializer, so this implies that a large
number of OverscrollActionsControllers are being initialized. This CL
adds some histograms to verify if this is the case.

Bug: 1102494
Change-Id: Id38c61781a8293a033a77df37227717e2ce7fe56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2391232
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#806596}
GitOrigin-RevId: cc7fafa82abf7152073e0438b69f90c6e260c34c