Remove ability for a URLRequest to use a separate NetworkDelegate.
The URLRequest constructor could accept a NetworkDelegate argument
which would override the URLRequestContext's NetworkDelegate. However,
URLRequest's constructor was private, and only URLRequestContext could
call it, and URLRequestContext didn't have a method to pass a
NetworkDelegate other than its own to URLRequest's constructor. As a
result, not only was nothing taking advantage of this ability, but
nothing *could* take advantage of it.

To the extent of my knowledge, this ability has never been used, yet
the per-request NetworkDelegate argument was wired through the entire
URLRequestJob creation stack.

Since this CL is already touching all URLRequestInterceptors, it
also updates the URLRequestInterceptor API to return a unique_ptr.

TBR= yuweih@chromium.org

Bug: 1126597
Change-Id: I5f0fade639e06a2edc8f514853262af3afa6e110
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2401854
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Monica Basta <msalama@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807048}
GitOrigin-RevId: fd97885a82707967300c8ab0a2058d79730d8bac