[iOS] Allow Disabling New Context Menu Delete Bookmark Action
In a previous CL, I've added logic to disable the Edit and Move actions
in the new Context Menus based on the enterprise policy's setting, but
forgot about the new Delete action. Bookmarks' deletion is also
controlled by this policy, so in this CL we'll be disabling that action
as well, based on the policy.

Bug: 1128028
Change-Id: I9f7cf4e8fa9e0889657d73a80ce1036d45d46511
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2410308
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807108}
GitOrigin-RevId: 0bf083888a60330f3a53b72663c7c6b382ede9e2