[iOS] Stop accessing pasteboard on startup in omnibox
Accessing the pasteboard on startup seems to cause hangs in the
pasteboard machinery often. These hangs are in system code, but we can
work around them by not accessing the pasteboard on startup.

For the location bar, this is relatively simple because we present the
edit menu ourselves. For the omnibox view controller, the edit menu
is presented by the system. To know the pasteboard state before the edit
menu is presented, we check the state when editing starts on the omnibox
text field.

Bug: 1121957
Change-Id: Icadb1710b2ea3684fe4250d556a3bd355339a0ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2414270
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807535}
GitOrigin-RevId: af3f29e537541e78dace8690ca74dd82a2dbe7c6