[iOS][test runner] Make DeviceTestRunner run xctest through xctest arg.
This change makes DeviceTestRunner also run target as XCTest by checking
xctest arg. Before this change, DeviceTestRunner uses solely the
existence of ".xctest" to determine if the target should run as XCTest.
This can cause trouble when targets build .xctest in but do not wish to
run as it.

The goal is to only use xctest arg to determine if it should run as
XCTest. These follow up CLs are needed to achieve the goal:
- Change device XCTest unit test configs in multiple repos.
- Remove the logic of using xctest_path to determine if run with xctest,
add a check to ensure xctest path is there when xctest is true.

Bug: 1085603
Change-Id: I07e5f21467791295e8ecc5f6246e58fd696d5e4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2412970
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807608}
GitOrigin-RevId: 0c71cc06dd5e71403898ca4d83084896bb849b13