Remove updatesCollectionView from grid view controller.
In order to allow the tab thumbnails to be up to date when the thumb
strip is visible, the grid's collection view needs to be updated even
when other view controllers are obscuring it. When selecting a new item
from the collection view (blue highlight around the thumbnail), the item
that is currently selected needs to be deselected to prevent multiple
simultaneous selection.

Bug: 1127604
Change-Id: Ic1e4a2dd4d0f2c269031b9d27b47b38c56b07a40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2410494
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Roberto Moura <mouraroberto@google.com>
Auto-Submit: Roberto Moura <mouraroberto@google.com>
Cr-Commit-Position: refs/heads/master@{#809263}
GitOrigin-RevId: 28310fb55130113c2009d88431728a4b3ec2ccf1