[ios] Move snapshot image during drag and drop
Previously, there was only one snapshot cache so the image did
not need to be moved during drag and drop. Recently, there
was a change to have a snapshot cache per browser.

This CL moves the snapshot image from the source cache to the
destination cache when the corresponding WebState is moved.

Bug: 1132306
Change-Id: I30599d17e33f9610adc7d715675bcbf829eadf8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2438312
Commit-Queue: edchin <edchin@chromium.org>
Auto-Submit: edchin <edchin@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811943}
GitOrigin-RevId: ee2de44e07e646ab2d21c4a3e5d1ae4fa29b0b3d