Remove updatesCollectionView from grid view controller.
In order to allow the tab thumbnails to be up to date when the thumb
strip is visible, the grid's collection view needs to be updated even
when other view controllers are obscuring it. When selecting a new item
from the collection view (blue highlight around the thumbnail), the item
that is currently selected needs to be deselected to prevent multiple
simultaneous selection.

Bug: 1127604
Change-Id: Ice919b9b7c068d381d667b0ac0ad82c41abfed35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2426590
Commit-Queue: Roberto Moura <mouraroberto@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Roberto Moura <mouraroberto@google.com>
Cr-Commit-Position: refs/heads/master@{#811968}
GitOrigin-RevId: 02611dfd4d5ef88444399a5adc965520d1a6231b