[ios] Don't resize overlay presentation container if insignificant
Upon layout, the container view of the Infobar banner view is resized
to match the size of the banner. But then a subsequently triggered
relayout may have a new banner frame size that is small enough to cause
an infinite loop. Thus, container resizing is only done if the height
and width of the new calculated banner frame are different by more
than 0.01.

Bug: 1129761
Change-Id: I0b397c63763c392e021bcf54fb7c7a2df8f872eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2432661
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#814502}
GitOrigin-RevId: 7c051ee54d889fdc605f3718a15f8b01413cf925