[ios] Fix recording of "Session.*" metrics
With recent change to Multi-Window, the metrics were recorded per
window instead of being recorded per BrowserState. Moreover, only
the currently displayed window's metrics were sent when the app
was backgrounded.

To fix this, add a new base::SupportsUserData::Data object that is
attached to the ChromeBrowserState. This object, SessionMetrics,
stores the session metrics which are updated from events happening
in all of the windows.

In preparation of adding metrics that are only recorded for the
off-the-record ChromeBrowserState, pass a set of flags to the
method recording the session metrics to choose exactly which one
should be recorded.

Add a call to SessionMetrics::RecordAndClearSessionMetrics() for
both the main and off-the-record ChromeBrowserState on application
background event, and when the off-the-record ChromeBrowserState is
destroyed.

Bug: none
Change-Id: Idee8ef1a1fb0be7153d4ceb78d0a795d589a9848
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2450052
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#814730}
GitOrigin-RevId: a1e355fa5a2d12b9f9f813462d297431d8045331