[iOS] Thumb Strip: tab switcher button uses thumb strip transitions
When the thumb strip is in hidden or peeked state, the tab switcher
button will transition the thumb strip to revealed state.

This CL also adds a public method to ViewRevealingVerticalPanHandler to
set the state externally.

Finally, the CL does not support tapping on the tab switcher button when
the thumb strip is in revealed state. This is because the design has the
BVC covered in a shim in this state. This hasn't been implemented yet,
but the tab switcher button should be inaccessible.

Change-Id: Ic44ddd8dc676fbdeb787bc75c422f332a9405574
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2455648
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#815226}
GitOrigin-RevId: 2850e6e4e67be6aa03575964d41122ec1fa08654