[iOS] Adding DCHECK to all user signin coordinators
The following coordinators should only be used when the user is signed
in:
 + AddAccountSigninCoordinator
 + AdvancedSettingsSigninCoordinator
 + TrustedVaultReauthenticationCoordinator

ConsistencyPromoSigninCoordinator should only be signed in when the user
is not signed in.

This patch adds DCHECK to enforce that.

Related to: crrev.com/c/2461345

Change-Id: I83282a15833f9450f0458793654d4a2da0ad0d25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2464274
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#817041}
GitOrigin-RevId: 66a5c48216f7d5b425d5d95b7635a9826848148d