Add a New Tab button (plus sign cell) to the end of the Thumb Strip.
The plus sign cell is a cell added to the end of the collection view of
the Tab Grid / Thumb Strip that opens a new tab when tapped. It has the
same shape of the thumbnail cells but has a plus sign in its center. A
new class was created for that: PlusSignCell.

The plus sign cell is always the last element of the collection view.

Screenshot: http://screen/49Y5RU8dPwDVXkW.png

Bug: 1127604, 1137983, 1137986
Change-Id: Ib1d9df1c22e6d3c2ca1340317dc7a39104275663
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2431668
Commit-Queue: Roberto Moura <mouraroberto@google.com>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Roberto Moura <mouraroberto@google.com>
Cr-Commit-Position: refs/heads/master@{#817498}
GitOrigin-RevId: b9c90c649ea875144ffdc9dd54c40f6d6447e70f