Drop SyncUserSettings::EncryptEverything()
This API was needed in the Directory implementation of Nigori as part
of the process of enabling a custom sync passphrase. Nowadays this is
taken care by SetEncryptionPassphrase(), so this CL removes it. It had
three remaining callers: the Android and iOS ones are easy to adapt; the
third is the implementation of the SyncBrowserProxy.setSyncEncryption()
Javascript API, called from the sync settings WebUI page.

setSyncEncryption() has poorly-defined semantics (crbug.com/1139060),
so it requires a few more changes. In short, instead of receiving only
what's needed ("Is a new passphrase being set? Which one?"), it takes
an entire SycPrefs [1] object, with the would-be parameters encoded by
callers in the setNewPassphrase/passphrase/encryptAllData fields. This
CL takes a first step in cleaning this up. The encryptAllData field
is no longer set in the JS, nor parsed in the backend, since it was
only used to decide whether to call EncryptEverything(). This already
requires that we update some JS tests: they used to verify the entire
prefs object, now they must verify only the necessary fields.

Next CLs will modify the signature of setSyncEncryption() to take only
what's needed.

[1] https://source.chromium.org/chromium/chromium/src/+/97d49f526544f71dcef5b3ed21f03c507014cff7:chrome/browser/resources/settings/people_page/sync_browser_proxy.js;l=94

TBR=jlebel@chromium.org

Bug: 1033040
Change-Id: Ibe508f60dc7d16096bddf808b31a0fe8046010d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2476637
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#817944}
GitOrigin-RevId: 6904cc7fdef0f3d9d64b5b47c5fe6f0edbf55430