Automated refactor: Migrate sync code to new gmock macros
One problem with the old MOCK_METHODN/MOCK_CONST_METHODN gmock macros
is that they never specify whether the mocked method is an override or
not. The consequence is that sometimes, methods are removed from sync
interfaces but remain in the mock implementations. Worse, they can even
even remain in test code if the test checks something of the form
EXPECT_CALL().Times(0). This CL uses a script to migrate all the sync
code to the modern MOCK_METHOD macro.

For convenience, patchset 1 is a pure application of the script, marking
every method as 'override'. Patchset 2 fixes compilation errors, i.e.
either removes leftover methods, or removes the override annotation if
the method was indeed intended not be an override.

Bug: 1139379
Change-Id: Ic5178057afed6d809d52a0d9c8c5d881acfe20ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2480762
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#818398}
GitOrigin-RevId: 5c5ae50fcb3768b55593a49baaca6b742aa62900