Change Tab Grid's bottom inset when Thumb Strip is enabled.
Move kBVCHeightTabGrid (height of the BVC that remains visible when the
tab grid is visible) to the grid constants file.

Correct the Thumb Strip Layout top section inset value that was wrong.
Increase the Grid Layout bottom section inset to account for the BVC on
the bottom of the screen.
Correct the Tab Grid's bottom content inset by increasing it by the same
amount that the top content inset was decreased.

Bug: 1127604
Change-Id: I843d74f7ade7272f032211162a0bca9054602dc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2479467
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Roberto Moura <mouraroberto@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#818423}
GitOrigin-RevId: 1c9849d41e6ea8a1c28b797e91ac48129fd38b76