[ios] Set `target_environment` gn variable in setup-gn.py
In preparation of supporting arm64 simulator build (for Apple Silicon
macs), we need to decouple the fact that a build is a simulator build
from the `target_cpu`.

This was done by introducing a `target_environment` variable that can
be set by the user (the default value is set using the same heuristic
based on value of `target_cpu` if unset).

Configure setup-gn.py to be explicit about setting the variable.

Bug: 1138425
Change-Id: Ic8f84ef260972e542ea456d6e85e43080e81da36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2489984
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#819768}
GitOrigin-RevId: 2cb632552e547e9750e2f9d84d059696448f3152