Schedule discarded session cleanup on the app initialization.
the didDiscardSceneSessions callback can be called when the browser
state is not yet created, or it can be called when the browsing data
remover service is not available. Once the callback is received mark
the sessions for a later time removal by saving them on a plist on disk.
And at the app initialization after the browser state its keyed services
are initialized. schedule the cleanup for these sessions.

Bug: 1136761
Change-Id: Ia5190866d4af7a041f0e5507fb7a88ea0ace9901
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2490788
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#820022}
GitOrigin-RevId: c9818bdf1b958300936ded1ec230416fdec1fe7c