[iOS] Reenable OverscrollActions tests on device
The tests now pass after
https://chromium-review.googlesource.com/c/chromium/src/+/2446782. See
below for a hypothesis on the root causes.

The root cause of the issue is a unclear, but seems to be as follows:
The overscroll state is supposed to change twice, once to
STARTED_PULLING after the |-scrollViewDidScroll| call, and once to
ACTION_READY after the contentOffset is updated. This happens
on simulator, but on device, the first change actually causes the
overscroll state to change to NO_PULL_STARTED.

This seems to be because on device, the initial contentOffset is 0,
while on simulator, it is ~ -2.46E-314, which compares less than 0 in
|OverscrollActionsController scrollViewDidScroll|. To fix this, I
changed the content offset in the test manually to a small number less
than 0.

Bug: 944599
Change-Id: I0d062e6ab9d20aca0b85810375b0081186fef48a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2390381
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/master@{#820216}
GitOrigin-RevId: 6de57b8d8f67f6c61160558fb6c45ee918e8286d