[ios] Fix Info.plist and entitlements for "catalyst" environment
Move ios/chrome/app/resources/Info.plist keys specific to the
"device" or "simulator" environments to a separate file, and
add a file for "catalyst" specific keys. Update the build rule
to merge the correct .plist files.

Add a new entitlement file for key that are required for a build
targeting the "catalyst" environment and update the build rule
to merge it to the generic entitlements.

Bug: 1138425
Change-Id: I164aeadf66320885695fb38707f0b10c22e07158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2498487
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821148}
GitOrigin-RevId: 7007950cea94f5bdb04fe06ccb7bab78c43dd2de