[iOS] Fix tab strip views when using FullscreenProvider
When the FullscreenProvider is used, the layout of the browser views is
slightly different. Instead of having the web content views start from
the bottom of the toolbar, the web content extends all the way to the
top of the BrowserContainerViewController, behind the tab strip.

This means that the prior approach of making all the top background
views clear does not work when using FullscreenProvider. Instead, this
CL uses the following approach: set these background views to have
the same background color as the tab grid. As soon as the panning
gesture starts, make them clear so the tab grid shows through and the
animations appear correctly. At the same time, move the web content
views down so they don't obstruct the animations either.

Bug: 1094335
Change-Id: Ieea7d268e781da5e1ccd4c2c0ad3c0925407d133
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2489898
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#821200}
GitOrigin-RevId: c756c75c4411842b47f36fa3fb50693df0da6948