[bookmarks] Refactor tests for TitledUrlIndex
This addresses some leftovers of the old refactoring effort in
https://codereview.chromium.org/2569333003.

The unittest file is now renamed to match the name of the class under
test, which is TitledUrlIndex. Some mechanical refactoring is included
such that the unit-tests don't involve BookmarkModel, which is a
higher-level layer.

A few tests that actually do exercise the integration with
BookmarkModel are now moved to BookmarkModelTest.

Change-Id: Icf791ce0f88647c6305a4f460d2398cfb63bb155
Bug: 1129524
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2498544
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Eric Noyau <noyau@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821482}
GitOrigin-RevId: ecb93f4dce5da17cfc0543a4f2d2a904e41592cb