Fix base::Bind/Callback -> Once/Repeating for regressed directories
Use of base::Bind and base::Callback is deprecated in favor of the more
explicit Once/Repeating versions. A presubmit enforces this, except on
a set of files and directories that have not yet been converted. This
list, however, has not always been updated when a directory is migrated.
Unfortunately, a few such directories have since regressed.

This cl fixes those regressions and removes the legacy exemption for the
directories to prevent any future regressions.

Bug: 1141533, 1007678, 1007729, 1007733, 1007793
Change-Id: I5efa002606ebacb4efa9db14016cc96f3ac72fb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2500563
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#821535}
GitOrigin-RevId: 536193dae7961d131d45182a2cdc16d3d55770d6