Add and report real-time extension request
Record Profile path when a profile's extension request pending list
is updated.

And add the all profile paths to the report with BrowserReportGenerator
for extension request partial report. the profile path list is reset
when the report is generated.

One change in ProfileReportGenerator is that we still upload report
even the extension request list is empty. So that server side could
get real-time request removal event too.

Bug: 1137964
Change-Id: Ib2789430813b45bc7256a6ef6f18dac9288f6c51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2518896
Commit-Queue: Owen Min <zmin@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#824491}
GitOrigin-RevId: 6a169b567bdc5aef3287a9a8f9138fb274415da7