Show a plus sign button when the Thumb Strip's plus sign cell is hidden.
Create a plus sign button that sticks to right of the thumb strip. It
has a plus sign and a transparency gradient and is shown when the plus
sign cell is not visible.
This is done so that the user can open a new tab even if they haven't
scrolled to the end of the collection view. When the user scrolls the
collection view or when the toolbars are shown, check whether the user
has scrolled past the end of the scroll view and, if so, hide the plus
sign button. Show it otherwise.

Rename one of the LayoutSwitcherStates from "Full" to "Grid".

Add the constats that specify the new thumb strip plus sign button's
dimensions to the grid constants file.

Add a |didChangeLastItemVisibility| method to the grid view controller
delegate. Create a CGFloat property |fractionVisibleOflastItem| inside
the grid view controller, which indicates the visibility fraction of the
last item. Only call |didChangeLastItemVisibility| on the delegate if
the value of |fractionVisibleOflastItem| changes.

Bug: 1094335
Change-Id: I056d6d0ece4eb22f84e4904208bd0e83d9b93389
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2485167
Commit-Queue: Roberto Moura <mouraroberto@google.com>
Auto-Submit: Roberto Moura <mouraroberto@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#824659}
GitOrigin-RevId: 327c6845a617af1937953522c7b720be19ee4c6e