[iOS] Add CHECK for BVC layout crash
In some cases a crash is happening when building the BVC view hierarchy
(see the linked bug); this is due to UIKit throwing an exception when
a constraint is created between views that aren't in the same view tree.

This CL just adds an explicit CHECK for the views that were triggering
this crash, so that we can get more information about the state of these
views when this occurs.

Bug: 1136765
Change-Id: I9d9c0431663aaa26df38bac1b66c476938320d7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2510213
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#825440}
GitOrigin-RevId: be085ab007a8c19f162a031531f36fe001f5bf21