Sync: Wait for policies to be loaded before starting sync engine
Inject the policy::PolicyService in to syncer::StartupController and
wait for OnFirstPoliciesLoaded to be triggered before attempting to
start the sync engine. The wait is subjected to a timeout if it
takes too long to load the first chrome policies. In that case, we
assume that policies are loaded and try and start the sync engine.

Bug: 1129065
Change-Id: Ic209a607343a916df38b22b3227d930fecc44a74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2508814
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Commit-Queue: Yann Dago <ydago@chromium.org>
Cr-Commit-Position: refs/heads/master@{#826061}
GitOrigin-RevId: 413d39d155e3515b22157988264bb1c10fa2af14