[Siri Shortcuts] Donate Search in Chrome intent on query
Donates the Search In Chrome intent to SiriKit every time a search is
performed in the Omnibox. The donation is performed on the IO thread to
avoid blocking the UI thread. This also updates the intents definition
file to make the Search in Chrome intent available for Siri
suggestions, which was previously disabled and prevented intent
donation.

Bug: 1146116
Change-Id: I6f7beb004a7f38d3a6297c8e7c171675fef2277c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2521202
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#826305}
GitOrigin-RevId: 48d9345a7eeafe74ec43e3b440498ccc6a007603