[iOS CBCM] Enable CBCM at 100% stable
Enables CBCM by default on Chrome branded build (the
--enable-chrome-browser-cloud-management command-line switch is still
needed for Chromium builds).

Because the logic for checking whether CBCM is enabled or not is now
back to being the same on all platforms, this change also removes the
IsEnabled delegate method and brings the check back to the core
CBCMController class. As a result, the command switch declaration was
moved to components so it can be shared between platforms.

Bug: 1111435, 1097161
Change-Id: I5f15f06885b4deb3974616667273706931df6ad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2522981
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#826315}
GitOrigin-RevId: 4e6ac1e917fa4fd732a7d6a15b741948225b8506