[iOS Policy] Improve chrome://policy EG test checks
Some existing policy tests load the chrome://policy page and ensure that
it loaded correctly by checking for the checkbox label at the top of the
page. This doesn't help catch Web UI issues that happen when displaying
the policy table further down the page.

This CL changes the selector so it looks for the policy table header,
which should help catch more WebUI problems in the future.

Bug: 1147191
Change-Id: Ib6c0a7c13d689d12f62eca5c88873bb31e749891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2530173
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#826328}
GitOrigin-RevId: 4489ae919dffdd4a91362a6444c005fccdb3eb70