[Autofill] Compile default parsing patterns into binary.
The parsing patterns currently in autofill_regex_constants.{h,cc) are
about to be migrated to regex_patterns.json. This JSON file is loaded
asynchronously by the component updater.

However, tests need to load the JSON synchronously to avoid race
conditions. This CL compiles the JSON file to C++, so that the default
patterns are available from the start.

With this CL, the JSON is not loaded at all during runtime. Enabling
the component updater is left for a follow-up CL.

The growth of the binary size will be compensated when
autofill_regex_constants.{cc,h} are removed.

Bug: 1147608
Change-Id: Ibbdc9161ae253a80917cc7dd6154342e8fd7ed47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2532562
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827431}
GitOrigin-RevId: 88b7ca994a60e3692fc3face23ced98d119822e1