[ios] Fixes test failures in ContentSuggestionsTestCase.
Multiple content suggestions tests load pages with the same title, and
tab grid cells appear to stick around from test to test. They are hidden
and in the tab grid behind the BVC, so they can't be seen by the user,
but they are visible to EG matchers. EG tests were failing because the
same page title was shared by multiple tests, and the second test was
seeing duplicate labels between the NTP entry it was trying to tap and a
hidden tab grid cell from the previous test. This was not turning the
bots red because the test passed on retry, when run in isolation.

This CL fixes the test failures by adding grey_sufficientlyVisible() to
the matcher, so that it no longer matches against hidden tab grid cells.

BUG=1148938

Change-Id: If0d8b958cace5aa34dfe5e2ed854620b5adc51ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2538116
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827764}
GitOrigin-RevId: 46cfa63a87aac2f9acf056f0cf0adb59d3f651dd