[ios] Assert that the version of Xcode is high enough
The new Widget requires Xcode 12.0 or higher to build, so assert
that the version of Xcode used for the build is recent enough.

Update the documentation to recommend the same version of Xcode
as enforced by the code, and add tips on steps recommended to
perform when updating Xcode.

Bug: none
Change-Id: I8b8fbf3d53ef2ab98c2407e28e46ce4321cd72f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2527175
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#827785}
GitOrigin-RevId: 7080fe3b59e1b6d70fa2d49e570543617c7d9bf2