[ios] Speculative fix for testAppLaunchManagerForceRelaunchByKilling.
On the bots, these tests appear to fail because they can't find the
Restore infobar after a crash. This CL changes the test to wait for the
infobar to become visible, rather than doing a single check, in case the
test runs so quickly that the infobar hasn't yet had a chance to appear.

BUG=1149370

Change-Id: I8b65c1796408e8151582398d4b2c3c12534c209b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2540426
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#827817}
GitOrigin-RevId: 27ccfb23e489ec6d566a49b4fedefa028d9fe359