[ios] Avoid passing pointer to BookmarkNode* to blocks
Passing pointer to C++ objects to blocks is dangerous as it can lead
to use after free. Change BookmarkHomeViewController implementation
to instead pass GURL or node ids (recovering the node by iterating
over the bookmarks).

This causes edit/move/remove operation to do a linear search of the
BookmarkNode* in the BookmarkModel which can cause those operations
to become slower if the user has a large number of bookmarks. If it
is a problem, then BookmarkHomeViewController should probably use a
proxy object instead of directly referencing the BookmarkNode* with
the proxy listening for node deletion.

Bug: 1113618
Change-Id: I15cb714f15950194ed7a0cf821a3ff36cb52417b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2544705
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#828702}
GitOrigin-RevId: 2c78005c796aa39756aa1257175afb9ea194e628