Add source port value to CanonicalCookie
Save the port value of the response that set a cookie in
CanonicalCookie. This will first be used for metrics leading to a
potential cookie spec change.

This value is connected up to the persistent store which was
modified to support it in an earlier CL: https://crrev.com/c/2505468

Bug: 1148013, 1141135, 1031721
Change-Id: Ic849bb71403a05ab31fd37d3acf37453dac8ebb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2537938
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#828797}
GitOrigin-RevId: 21385a8c560b2e541301ab65b5419668ba38b601