[iOS] Don't update item URL if it is the one being restored
This CL changes the update of the item's URL when the WebView URL is
changed. If the new WebView URL is different from the URL of the item
but is the URL associated with the session restoration of this URL,
the URL of the item is no longer updated.

Fixed: 1139289
Change-Id: Id3b31bfedeb6e84bc0527f642fa424e7f81381dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2549620
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#829605}
GitOrigin-RevId: ba99071717a0686d97eee5f5b96df6dc09e1116c