[iOS][Thumb Strip] Keep BVC's position fixed when switching to Incognito
Before, when switching to incognito using the Thumb Strip, the BVC ended
up having a frame.y = 36 (or some other non-zero constant). It looks
like the BVC is currently positioned by frame, and my guess is that
somehow, the position of some other element + constraints caused the
BVC's frame to change after it was set. However, the BVC's frame could
be set by constraint, as it should always be that value.

Bug: 1094335
Change-Id: Iedeedadf7ac7c7e842ebe4af02d8e62ed3b6a7c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2549947
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#830146}
GitOrigin-RevId: 99a42e2d128ae0993f574294cfbab2a1f1f8dad4