Fix typo in policy_test_cases.json
This CL is part of a larger refactoring/clean-up of
policy_test_cases.json.

The code used to only looks for the field "policy_pref_mapping_test",
even though the template defined to use "policy_pref_mapping_tests".
Some test were actually using the plural version, which caused them to
be skipped.
Even though the singular version of the field name is used more often,
I've decided to update all occurrences to the plural version since it
should contain a list and already caused confusion multiple times
where the field value was an object instead (see next CL).

This is just a trivial global rename, except for the test case for
DeviceLoginScreenPrivacyScreenEnabled, which had to be disabled for
now.

Bug: 1151251
Change-Id: I454c86ca8d91abe260606ff1f0c86438c4cdebb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2552407
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Alexander Hendrich <hendrich@chromium.org>
Cr-Commit-Position: refs/heads/master@{#830181}
GitOrigin-RevId: 71376fe68cb914e7974187f1b19db878ebc509f2