Refactor pref type to allow for more values in the future
This CL is part of a larger refactoring/clean-up of
policy_test_cases.json.

Switching from a single "local_state" boolean to a "location", which can
have one of multiple values. This CL is not changing any behavior, but
acts as preparation for the next CLs, which will add more potential
locations ("signinProfile" + "crosSettings").

Bug: 1151251
Change-Id: Iaa7e6e39251d7333cbad053d756b4eda78150d34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2550558
Commit-Queue: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#830292}
GitOrigin-RevId: 488ecd9cfd232c52d4bb042381452dc0d5e9fa2c