[iOS][OpenIn] Refactors EG2 tests
This CL refactors EG2 tests related to the open in toolbar.
The toolbar and the display of the activity view is now tested
on a PDF file and a PNG image.
It also Tests that the open in bar is correctly displayed
when a compatible URL is loaded and tapping on the web view
makes it appear or disappear.

Bug: 1124310, 1146303, 1135805
Change-Id: I57f962a31a7e34a02ae37fb72bf649abb6fb74fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2532004
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#830570}
GitOrigin-RevId: b88b3f3521f27a533df12cc7d056fa9915ab5766