[iOS] Set FirstResponderSendAction as default
Remove experiment flag, and set enabled behavior as default.

Keep old behavior when asking the NTP and WebView to become first
responder. This is done checking if the first responder is a view, and
asking if not. Because in the previous logic for GetFirstResponder() the
search was made in the view hierarchy, ignoring view controllers.

The experiment enabled without the check added in this CL could explain
issues like crbug.com/1150498.

Bug: 816427, 1150498
Change-Id: Ib9ac17b59da8635d00beff071aee4bd6afeedb78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2495976
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Javier Flores <javierrobles@chromium.org>
Auto-Submit: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#830632}
GitOrigin-RevId: 2aa1a2f6694992bce4ae4392cac12f9109139c9e