[iOS] Unfocus omnibox when recreating NTP
This CL unfocusing the omnibox when the NTP is created. This is needed
as the NTP is re-created when the browsing data are cleared. This can
happen while the omnibox is focused on multiwindow.
Unfocusing the omnibox ensure that the NTP state (i.e. scrolled down)
is matching the omnibox state.

Bug: 1139191
Change-Id: I07d050bc1c2d90178bf80786150fd2cbc1d4e4f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2550044
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#831000}
GitOrigin-RevId: 7bf9d04f29fbea42526b19da9e41c619f6f3350e